Updated: MP3TagChecker

+++ apps/Mezzmo-ejb/src/be/home/mezzmo/domain/model/MGOFileArtistTO.java
@@ -5,14 +5,14 @@ package be.home.mezzmo.domain.model;
  */
 public class MGOFileArtistTO {
 
-    private int ID;
+    private Long ID;
     private String artist;
 
-    public int getID() {
+    public Long getID() {
         return ID;
     }
 
-    public void setID(int ID) {
+    public void setID(Long ID) {
         this.ID = ID;
     }
